requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.12
asttokens==2.2.0
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
dill==0.3.6
Django==4.1.1
django-allauth==0.54.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-filter==22.1
django-import-export==3.2.0
django-jazzmin==2.6.0
django-parsley==0.7
django-redis==5.2.0
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
factory-boy==3.2.1
Faker==17.0.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipython==8.7.0
isort==5.10.1
jedi==0.18.2
kombu==5.2.4
lazy-object-proxy==1.8.0
MarkupPy==1.14
matplotlib-inline==0.1.6
mccabe==0.7.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.33
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.7.0
pylint==2.15.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytest==7.2.1
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
razorpay==1.3.0
redis==4.5.5
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.2
stripe==5.2.0
tablib==3.4.0
tomli==2.0.1
tomlkit==0.11.6
traitlets==5.6.0
typing-extensions==4.4.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0


django==5.0.4
celery==5.3.6
django-jazzmin==2.6.1
django-extensions==3.2.3
django-filter==24.2
django-import-export==3.3.7
django-widget-tweaks==1.5.0
django-allauth==0.54.0
razorpay==1.4.2
Pillow==10.3.0